requirements.txt
# P0 — FCA CASS 7.15 / CASS 15
# banxe-emi-stack runtime dependencies

# FastAPI REST API Layer (IL-046)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
PyJWT>=2.8.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.0

# Midaz CBS client
httpx>=0.27.0

# ClickHouse audit trail
clickhouse-driver>=0.2.9

# CAMT.053 / MT940 bank statement parsing (Phase 2)
# bankstatementparser>=1.0.0
# mt940>=4.30.0

# dbt transformations
dbt-clickhouse>=1.8.0

# FIN060 PDF generation
weasyprint>=62.3

# FX rates (Frankfurter self-hosted)
# (uses httpx above, no additional dep)

# Redis VelocityTracker (IL-048)
redis>=5.0.0

# RabbitMQ event bus (IL-053)
pika>=1.3.0

# PostgreSQL config store (IL-053)
psycopg2-binary>=2.9.9

# YAML config store
pyyaml>=6.0

# Dev / test
pytest>=8.0
pytest-asyncio>=0.23
fakeredis>=2.21.0
